requirements.txt
winsdk
pillow
customtkinter
asyncio
numpy